feat: implement user profile page
